refactor: 닉네임 입력 컴포넌트 분리 및 공백 허용 로직 수정
