+++ src/app/nicknamesetting/NicknamePage.module.scss
@@ -15,36 +15,3 @@
     @include text-style(24, $black, bold);
   }
 }
-.input-group {
-  @include column-flexbox(center, start, 0.6rem);
-
-  width: 100%;
-
-  .title {
-    @include text-style(18, $black, bolder);
-  }
-
-  .description {
-    @include text-style(14, $black, light);
-
-    margin-bottom: 0.4rem;
-  }
-
-  .input {
-    @include text-style(14);
-
-    transition: $base-transition;
-    width: 100%;
-    border-radius: 0.6rem;
-    border: 0.05rem solid $gray;
-    padding: 1.4rem 1.5rem;
-    &::placeholder {
-      @include text-style(14, $gray, light);
-    }
-
-    &:focus {
-      border-color: $primary;
-      outline: none;
-    }
-  }
-}
